[Fix]#43 - zigzag 받아오기 버그 해결

+++ src/main/java/nunu/orderCount/domain/order/model/dto/response/ResponseOrderUpdateDto.java
@@ -6,6 +6,6 @@ import lombok.RequiredArgsConstructor;
 @Getter
 @RequiredArgsConstructor
 public class ResponseOrderUpdateDto {
-    private final Long newOrderCount;
-    private final Long changeStatusOrderCount;
+    private final Integer newOrderCount;
+    private final Integer changeStatusOrderCount;
 }
